Define the Object type for shit with ObjectRead and ObjectWrite utils (#4)
## Changelog 

This pull request introduces a new object model and associated
functionalities for handling different types of objects in the
repository. The most important changes include the addition of the
`ShitBlob` type, the implementation of the `Object` interface, the
creation of utility functions for file operations, and the corresponding
unit tests.

* Implemented the `Object` interface, including methods for reading and
writing objects, and an `ObjectFactory` function to create objects based
on type.
* Added the `ShitBlob` type with methods to initialize, serialize, and
deserialize blob data.


## Testing 
* Added unit tests for the new object model, including tests for object
creation, reading, writing, and utility functions.