Support CommitMetadata SerDe operations (#10)
## Changelog 
Support commit metadata serialization and deserialization operations for
a git commit object.


I did not like the KLVM structure that was defined in
https://wyag.thb.lt/#orgbd44de2 so decided to use
https://github.com/go-git/go-git/blob/master/plumbing/object/object.go#L81
as reference for the SerDe operations. This seemed much cleaner as the
fields were defined explicitly.

This also means that #9 should be invalid now. We should be able to
close it out.

## Testing 
Added unit test to test out the functionality for Ser, De, Ser -> De, De
-> Ser